+++ assets/graveyard/blood_magic.rb
