+++ assets/graveyard/combat/beatdown.rb
