+++ assets/graveyard/combat/boss_band.rb
