+++ assets/graveyard/combat/boss_barkeep.rb
