+++ assets/graveyard/combat/boss_bouncer.rb
