+++ imports.rb
@@ -8,17 +8,11 @@ require_relative 'assets/ascii'
 require_relative 'assets/cheat_mode'
 require_relative 'assets/names'
 require_relative 'assets/phrases'
-require_relative 'combat/beatdown'
-require_relative 'combat/boss_band'
-require_relative 'combat/boss_barkeep'
-require_relative 'combat/boss_bouncer'
 require_relative 'display/display_info'
 require_relative 'display/display_ui'
-require_relative 'display/menu_boss'
 require_relative 'display/menu_games'
 require_relative 'display/menu_main'
 require_relative 'display/white_space'
-require_relative 'main/big_boss'
 require_relative 'main/attack_mode'
 require_relative 'main/escape_room'
 require_relative 'main/war_machine'
@@ -27,5 +21,4 @@ require_relative 'messages/invoice'
 require_relative 'messages/narrative'
 require_relative 'messages/war_chatter'
 require_relative 'powers/blackjack'
-require_relative 'powers/blood_magic'
 #----------------------------- IMPORTS ---------------------------------------->
